🗃Refactor: Cafe 클래스 생성
