+++ src/domain/beverage/service/BeverageService.ts
@@ -24,8 +24,6 @@ export class BeverageService {
   public getBeverageById(id: number): Beverage {
     const beverage = this.beverages.find(beverage => beverage?.id === id);
 
-    console.log(beverage);
-
     if (!beverage) {
       throw new Error();
     }
